+++ src/resources/views/livewire/parts/column-toggle.blade.php
@@ -3,7 +3,7 @@
     <x-beartropy-ui::button 
         @click="isOpenColumnToggle = ! isOpenColumnToggle"
         variant="{{ $yat_button_variant }}"
-        color="{{ $theme }}"
+        color="{{ $buttonThemeOverride ?? $theme }}"
         class="w-full"
     >
         <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-columns mr-2">
@@ -73,7 +73,7 @@
                                 id="{{ $column->key }}" 
                                 label="{{ $column->label }}"
                                 class="w-full py-3 md:py-1.5"
-                                color="{{ $theme }}"
+                                color="{{ $inputThemeOverride ?? $theme }}"
                                 sm
                             />
                         </div>