+++ src/resources/views/livewire/parts/global-search.blade.php
@@ -4,7 +4,7 @@
         placeholder="{{ucfirst(__('yat::yat.search'))}}" 
         clearable 
         class="w-full"
-        color="{{ $theme }}"
+        color="{{ $inputThemeOverride ?? $theme }}"
     >
         <x-slot:start>
             <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="gray" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-search">