+++ src/resources/views/livewire/parts/options.blade.php
@@ -3,7 +3,7 @@
     <!-- Toggle Button -->
     <x-beartropy-ui::button 
         variant="{{ $yat_button_variant }}"
-        color="{{ $theme }}"
+        color="{{ $buttonThemeOverride ?? $theme }}"
         @click="isOpenOptionsToggle = ! isOpenOptionsToggle"
         class="w-full"
     >