+++ src/resources/views/livewire/parts/pagination.blade.php
@@ -6,7 +6,13 @@
                     {{ucfirst(__('yat::yat.showing'))}} {{ $rows->firstItem() }} {{ucfirst(__('yat::yat.to'))}} {{ $rows->lastItem() }} {{ucfirst(__('yat::yat.of'))}} {{ $rows->total() }} {{ucfirst(__('yat::yat.entries'))}}
                 </div>
                 <div class="inline-flex">
-                    {{ $rows->links('yat::livewire.parts.pagination-actions', ['themeConfig' => $themeConfig]) }}
+                    @php
+                        $paginationThemeConfig = $themeConfig;
+                        if (!empty($buttonThemeOverride) && $buttonThemeOverride !== $theme) {
+                            $paginationThemeConfig = $this->getThemeConfig($buttonThemeOverride);
+                        }
+                    @endphp
+                    {{ $rows->links('yat::livewire.parts.pagination-actions', ['themeConfig' => $paginationThemeConfig]) }}
                 </div>
             </div>
         </div>