+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.10] - 2025-12-18
+
+### Changed
+- Updated defaults in `View` trait to fallback to `gray` theme.
+- Updated pagination components to respect `buttonThemeOverride`.
+
+### Added
+- Added `getThemeConfig` helper method to `View` trait.
+
 ## [2.3.9] - 2025-12-18
 
 ### Changed